requirements.txt
dj-database-url==3.1.2
django-cors-headers==4.9.0
django-phonenumber-field==8.4.0
Django==6.0.2
djangorestframework_simplejwt==5.5.1
djangorestframework==3.16.1
factory_boy==3.3.3
phonenumbers==9.0.2
pre-commit==4.5.1
psycopg2-binary==2.9.11
uuid_utils==0.14.1
